spmi: hisi-spmi-controller: manage the OF node reference in device initialization and cleanup
spmi_controller_probe() increments the refcount of an OF node, but does
not release it. Instead, call of_node_get() in spmi_controller_alloc()
and release it in spmi_ctrl_release() to avoid the reference leak. Also
remove the lines in spmi_pmic_arb_bus_init() and spmi_controller_probe()
where a pdev's of_node is stored after spmi_controller_alloc() is
called, since the node is already set with an incremented refcount in
spmi_controller_alloc(). Those lines do not overwrite the of_node since
they set the same node that the spmi_controller_alloc() does, but are
confusing at the very least.

This bug was found by an experimental verification tool that I am
developing.

Fixes: e562cf3aea3e ("spmi: hisi-spmi-controller: move driver from staging")
Signed-off-by: Joe Hattori <joe@pf.is.s.u-tokyo.ac.jp>
Link: https://lore.kernel.org/r/20241219013507.1939266-1-joe@pf.is.s.u-tokyo.ac.jp
Signed-off-by: Stephen Boyd <sboyd@kernel.org>